use int for y_oh in eval

+++ columbo/runners.py
@@ -129,7 +129,7 @@ def run_training(epochs:int=10, classifier: Union[str, Callable]=None, device=No
                 X, y = sample[0], sample[1]
                 X1 = X.unsqueeze(dim=1)
                 y_pred = model(X)
-                y_oh = torch.nn.functional.one_hot(y, num_classes=2).float()
+                y_oh = torch.nn.functional.one_hot(y, num_classes=2)
                 loss = loss_fn(y_pred.squeeze(), y_oh)
                 val_loss +=  loss.cpu().detach().numpy()
